pizza_app/requirements.txt
requests
pyhamcrest
xmltodict
pytest
mock
djangorestframework
django-cors-headers